dto files moved to their respective apis folder
