+++ src/application/application-roles/application-roles.service.ts
@@ -1,6 +1,6 @@
 import { HttpException, HttpStatus, Injectable, Logger } from '@nestjs/common';
 import { randomUUID } from 'crypto';
-import { RoleDto, UpdateRoleDto } from 'src/dto/application.dto';
+import { RoleDto, UpdateRoleDto } from 'src/application/application.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
 
 @Injectable()
