+++ src/key/key.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { KeyController } from './key.contoller';
 import { KeyService } from './key.service';
-import { generateKeyDTO, updateDTO } from 'src/dto/key.dto';
+import { generateKeyDTO, updateDTO } from 'src/key/key.dto';
 
 @Module({
   controllers: [KeyController],
